+++ riscv-runtime/src/input.rs
@@ -7,6 +7,14 @@ use powdr_riscv_syscalls::Syscall;
 use alloc::vec;
 use alloc::vec::Vec;
 
+pub fn get_prover_input(index: u32) -> u32 {
+    let mut value: u32;
+    unsafe {
+        asm!("ecall", lateout("a0") value, in("a0") index, in("t0") Syscall::Input as u32);
+    }
+    value
+}
+
 pub fn get_data(channel: u32, data: &mut [u32]) {
     for (i, d) in data.iter_mut().enumerate() {
         unsafe {
@@ -34,32 +42,4 @@ pub fn get_data_serde<T: DeserializeOwned>(channel: u32) -> T {
     let data: Vec<u8> = data.into_iter().map(|x| x as u8).collect();
 
     serde_cbor::from_slice(&data.as_slice()).unwrap()
-}
-
-const GOLDILOCKS: u64 = 0xffffffff00000001;
-
-/// Calls the low level Poseidon coprocessor in PIL, where
-/// the last 4 elements are the "cap"
-/// and the return value is placed in data[0:4].
-/// The safe version below also checks that each u64 element
-/// is less than the Goldilocks field.
-/// The unsafe version does not perform such checks.
-pub fn poseidon_gl(mut data: [u64; 12]) -> [u64; 4] {
-    for &n in data.iter() {
-        assert!(n < GOLDILOCKS);
-    }
-
-    unsafe {
-        asm!("ecall", in("a0") &mut data as *mut [u64; 12], in("t0") Syscall::PoseidonGL as u32);
-    };
-
-    [data[0], data[1], data[2], data[3]]
-}
-
-pub fn poseidon_gl_unsafe(mut data: [u64; 12]) -> [u64; 4] {
-    unsafe {
-        asm!("ecall", in("a0") &mut data as *mut [u64; 12], in("t0") Syscall::PoseidonGL as u32);
-    }
-
-    [data[0], data[1], data[2], data[3]]
 }
