+++ riscv-runtime/src/lib.rs
@@ -11,11 +11,10 @@ use core::panic::PanicInfo;
 
 use crate::fmt::print_str;
 
-use powdr_riscv_syscalls::Syscall;
-
 mod allocator;
-pub mod coprocessors;
 pub mod fmt;
+pub mod hash;
+pub mod input;
 
 #[panic_handler]
 unsafe fn panic(panic: &PanicInfo<'_>) -> ! {
@@ -31,15 +30,6 @@ unsafe fn panic(panic: &PanicInfo<'_>) -> ! {
 
     asm!("unimp");
     loop {}
-}
-
-#[inline]
-pub fn get_prover_input(index: u32) -> u32 {
-    let mut value: u32;
-    unsafe {
-        asm!("ecall", lateout("a0") value, in("a0") index, in("t0") Syscall::Input as u32);
-    }
-    value
 }
 
 extern "Rust" {
