+++ riscv/tests/riscv_data/poseidon_gl_via_coprocessor/src/lib.rs
@@ -1,6 +1,6 @@
 #![no_std]
 
-use powdr_riscv_runtime::coprocessors::{poseidon_gl, poseidon_gl_unsafe};
+use powdr_riscv_runtime::hash::{poseidon_gl, poseidon_gl_unsafe};
 
 #[no_mangle]
 fn main() {
