+++ riscv/tests/riscv_data/print/src/lib.rs
@@ -1,6 +1,7 @@
 #![no_std]
 
-use powdr_riscv_runtime::{get_prover_input, print};
+use powdr_riscv_runtime::input::get_prover_input;
+use powdr_riscv_runtime::print;
 
 #[no_mangle]
 pub fn main() {
