+++ riscv/tests/riscv_data/sum/src/lib.rs
@@ -3,7 +3,7 @@
 extern crate alloc;
 use alloc::vec::Vec;
 
-use powdr_riscv_runtime::get_prover_input;
+use powdr_riscv_runtime::input::get_prover_input;
 
 #[no_mangle]
 pub fn main() {
