+++ riscv/tests/riscv_data/sum_serde/src/lib.rs
@@ -3,7 +3,7 @@
 extern crate alloc;
 use alloc::vec::Vec;
 
-use powdr_riscv_runtime::{coprocessors::get_data_serde, get_prover_input};
+use powdr_riscv_runtime::input::{get_data_serde, get_prover_input};
 
 #[no_mangle]
 pub fn main() {
