+++ riscv/tests/riscv_data/vec_median/src/lib.rs
@@ -16,7 +16,8 @@
 extern crate alloc;
 
 use alloc::vec::Vec;
-use powdr_riscv_runtime::{get_prover_input, print};
+use powdr_riscv_runtime::input::get_prover_input;
+use powdr_riscv_runtime::print;
 
 #[no_mangle]
 fn main() {
